+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.5.0
 
 - changed "natural stamina regeneration" gamerule to simply add 1 stamina regeneration. The previous implementation prevented stamina regeneration from ever becoming negative.
+- fixed "offset_from_origin" config values not working correctly
+- stamina number display is now independent of the stamina bar
 
 # 2.4.0